# application/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "This API is designed for a Mechanic Shop Management System"
  version: "1.0.0"
host: "127.0.0.1:5000" # working on local host
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"

####********************************** Documenting the Customer Route******************************####
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      Enter your bearer token in the format: `Bearer <token>`
  ##========== Create Login for Customer (POST) ==========##
paths:
  /customers/login:
    post:
      tags:
      - Customers
      summary: "Login endpoint, (generates token)"
      description: "Validates customer credentials, generates and return a customer specific token."
      parameters:
      - in: "body"
        name: "body"
        description: "Login Credentials"
        required: true
        schema:
          $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NTE5OTAwOTcsImlhdCI6MTc1MTk4NjQ5Nywic3ViIjoiMSJ9.sEjSewn29J41DxWj5xaRBBgj_qq8JS228HhcBXAbfoU"
              message: "Successfully Logged In and the Customer name is: yashika , customer id is: 1"
              status: "Success"

  ##=======Documenting Create Customer Route (POST) =======##
  /customers:
    #endpoint
    post:
      #Post request
      tags:
      - Customers
      summary: "Endpoint to create/register a new customer."
      description: "Endpoint to create a new customer in our Mechanic shop system, emails must be unique."
      parameters:
      - in: "body"
        name: "body"
        description: "Information required to create a new Customer."
        required: true
        schema:
          $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Successfull Customer Creation"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              message: "New Customer details added successfully"
              id: 1
              name: Yashika
              email: yashika@email.com
              password: yashi
              address: 1st, southwest street
              phone: 111-222-1112
              salary: 80000.00

    ##===== Documenting Retrieve Customers using pagination (GET) ======##

    get:
      tags:
      - Customers
      summary: "Get list of customers (supports pagination)"
      description: |
        Retrieve a list of customers.  
        Supports pagination using `page` and `per_page` query parameters.
      parameters:
      - in: query
        name: page
        type: integer
        description: Page number
      - in: query
        name: per_page
        type: integer
        description: Number of items per page
      responses:
        '200':
          description: Successful retrieval of customers list
          schema:
            type: array
            items:
              $ref: "#/definitions/AllCustomersPaginationResponse"
          examples:
            application/json:
              id: 4
              name: Rosy
              email: rosy@gmail.com
              password: rosy
              address: 200, circle road
              phone: 574-212-2675
              salary: 11000

    ##=======Documenting Update Customer Route (PUT) =======##

    put:
      tags:
      - Customers
      summary: "Endpoint to update a customer."
      description: "Endpoint to update a customer, this is a token authenticated route, and the customer must be logged in to update their account."
      security:
      - bearerAuth: []
      parameters:
      - in: "body"
        name: "body"
        description: "Information required to update a Customer."
        required: true
        schema:
          $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Successfully Updated Customer"
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"
          examples:
            application/json:
              message: "Successfully Updated Customer id: 1"
              id: 1
              name: Yashika
              email: yashika@email.com
              password: yashi
              address: 1st, southwest street
              phone: 111-222-1112
              salary: 80000.00

    ##=======Documenting Delete Customer Route (Delete) =======##

    delete:
      tags:
      - Customers
      summary: "Enpoint to delete a customer"
      description: "Endpoint to delete a customer. Requires token authentication, customer must be logged in to delete their account"
      security:
      - bearerAuth: []
      responses:
        200:
          description: "Successfully Deleted Customer"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "Customer id: 3, Successfully deleted."

  ##=======Documenting Get Specific Customer Route (GET) =======##
  /customers/{customer_id}:
    get:
      tags:
      - Customers
      summary: "Get a customer by ID"
      description: "Customer details are generated based on the customer ID and return a specific customer."
      parameters:
      - in: path
        name: customer_id
        required: true
        schema:
          type: integer
        description: ID of the customer to retrieve
      responses:
        '200':
          description: Customer found
          schema:
            $ref: "#/definitions/SpecificCustomerResponse"
          examples:
            application/json:
              id: 4
              name: Rosy
              email: rosy@gmail.com
              password: rosy
              address: 200, circle road
              phone: 574-212-2675
              salary: 11000

  ####********************************** Documenting the Create Mechanic Route******************************####
  ##=======Documenting Create Mechanic Route (POST) =======##
  /mechanics:
    #endpoint
    post:
      #Post request
      tags:
      - Mechanics
      summary: "Endpoint to create/register a new Mechanic."
      description: "Endpoint to create a new mechanic in our Mechanic shop system, emails must be unique."
      parameters:
      - in: "body" ### passing the body parameter
        name: "body"
        description: "Information required to create a new Mechanic."
        required: true
        schema:
          $ref: "#/definitions/CreateMechanicPayload"
      responses:
        201:
          description: "Successfull Mechanic Creation"
          schema:
            $ref: "#/definitions/CreateMechanicResponse"
          examples:
            application/json:
              message: "New Mechanic details added successfully"
              id: 2
              name: Legha
              address: 2, South street
              email: legha@gmail.com
              phone: 897-543-6785

    ##=======Documenting Get all Mechanics Route (GET) =======##

    get:
      #Get request
      tags:
      - Mechanics
      summary: "Returns all Mechanics"
      description: "Endpoint to retrieve a list of all mechanics."
      responses:
        200:
          description: "Retrieved Mechanics Successfully"
          schema:
            $ref: "#/definitions/AllMechanicsResponse"

  ##=======Documenting Get Specific Mechanic Route (GET) ======##
  /mechanics/{mechanic_id}:
    get:
      tags:
      - Mechanics
      summary: "Get a mechanic by ID"
      description: "Mechanic details are generated based on the mechanic ID and return a specific mechanic details."
      parameters:
      - in: path ### passing the path parameter
        name: mechanic_id
        required: true
        schema:
          type: integer
        description: ID of the mechanic to retrieve
      responses:
        '200':
          description: Mechanic found
          schema:
            $ref: "#/definitions/SpecificMechanicResponse"
          examples:
            application/json:
              id: 5
              name: Meena
              email: meena@gmail.com
              address: 100, North street
              phone: 997-543-6785

    ##=======Documenting Update Mechanic Route (PUT) =======##

    put:
      tags:
      - Mechanics
      summary: "Endpoint to update a mechanic by ID."
      description: "Endpoint to update a mechanic, Mechanic details are generated based on the mechanic ID and return a Updated mechanic details."
      parameters:
      - in: path ### passing the path parameter
        name: mechanic_id
        required: true
        type: integer
        description: ID of the mechanic to update
      - in: body
        name: body
        required: true
        description: "Information required to update a Mechanic"
        schema:
          $ref: "#/definitions/UpdateMechanicPayload"
      responses:
        200:
          description: "Successfully Updated Mechanic"
          schema:
            $ref: "#/definitions/UpdateMechanicResponse"
          examples:
            application/json:
              message: "Successfully Updated Mechanic id: 2"
              mechanic:
              - id: 2
                name: Legha
                address: 2, South street
                email: legha@gmail.com
                phone: 897-543-6785

    ##=======Documenting Delete Mechanic Route (Delete) =======##

    delete:
      tags:
      - Mechanics
      summary: "Enpoint to delete a mechanic by ID"
      description: "Endpoint to delete a Mechanic. Mechanic details are generated based on the mechanic ID and return a successful deleted mechanic."
      parameters:
      - in: path
        name: mechanic_id
        required: true
        type: integer
        description: ID of the mechanic to delete
      responses:
        200:
          description: "Successfully Deleted mechanic"
          schema:
            $ref: "#/definitions/DeleteMechanicResponse"
          examples:
            application/json:
              message: "Mechanic id: 3, Successfully deleted."

  ##=======Most Serivce Tickets worked on by Mechanic List Route (GET) =======##
...
  /mechanics/search:
    get:
      tags:
      - Mechanics
      summary: "Search mechanics by name"
      description: |
        Search and retrieve a list of mechanics whose name contains the specified query string
        Example: `/search?name=mee`
      parameters:
      - in: query ### passing the path parameter
        name: name
        type: string
        required: true
        description: Substring to search in mechanic names
      responses:
        '200':
          description: "List of matching mechanics"
          schema:
            type: array
            items:
              $ref: "#/definitions/MechanicNameSearchResponse"
          examples:
            application/json:
            - id: 5
              name: Meena
              email: meena@gmail.com
              address: 100, North street
              phone: 997-543-6785

  ####********************************** Documenting the Service_ticket Route******************************###
  ##=======Documenting Create Service_ticket with mechanic Route (POST) =======##
...
definitions:
...
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"

    required:
    - email
    - password
  LoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"

  ## Create Customer Request and Responses
  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
        format: email
        description: Unique email address for the customer
      password:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: number
        format: float
        description: Monthly salary

    required:
    - name
    - email
    - password
    - address
    - phone
    - salary
  CreateCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"
      id:
        type: integer
      name:
        type: "string"
      email:
        type: "string"
        format: email
      password:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: number
        format: float

  ## Retrieve Customers using pagination Responses
  AllCustomersPaginationResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: number
        format: float

  ## Update Customer Request and Responses
  UpdateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
        format: email
        description: Unique email address for the customer
      password:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: number
        format: float
        description: Monthly salary

    required:
    - name
    - email
    - password
    - address
    - phone
    - salary
  UpdateCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"
      id:
        type: integer
      name:
        type: "string"
      email:
        type: "string"
        format: email
      password:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: number
        format: float

  ## Delete Customer Response
  DeleteCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"

  ## Specific Customer Responses
  SpecificCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      address:
        type: "string"
      phone:
        type: "string"
      salary:
        type: number
        format: float

    required:
    - id
    - name
    - email
    - password
    - address
    - phone
    - salary

  ####**********************************Definitions for Mechanic Routes******************************####

  ## Create Mechanic Request and Responses
  CreateMechanicPayload:
    type: "object"
    properties:
      name:
        type: "string"
      address:
        type: "string"
      email:
        type: "string"
        format: email
        description: Unique email address for the mechanic
      phone:
        type: "string"

    required:
    - name
    - address
    - email
    - phone
  CreateMechanicResponse:
    type: "object"
    properties:
      message:
        type: "string"
      id:
        type: integer
      name:
        type: "string"
      address:
        type: "string"
      email:
        type: "string"
        format: email
        description: Unique email address for the mechanic
      phone:
        type: "string"

  ## ALL Mechanics Responses
  AllMechanicsResponse:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: integer
        name:
          type: "string"
        address:
          type: "string"
        email:
          type: "string"
          format: email
          description: Unique email address for the mechanic
        phone:
          type: "string"

  ## Specific Mechanic Responses
  SpecificMechanicResponse:
    type: "object"
    properties:
      id:
        type: integer
      name:
        type: "string"
      address:
        type: "string"
      email:
        type: "string"
        format: email
        description: Unique email address for the mechanic
      phone:
        type: "string"

    required:
    - name
    - address
    - email
    - phone

  ## Update Mechanic Request and Responses
  UpdateMechanicPayload:
    type: "object"
    properties:
      name:
        type: "string"
      address:
        type: "string"
      email:
        type: "string"
        format: email
        description: Unique email address for the mechanic
      phone:
        type: "string"

    required:
    - name
    - address
    - email
    - phone
  UpdateMechanicResponse:
    type: "object"
    properties:
      message:
        type: "string"
      id:
        type: integer
      name:
        type: "string"
      address:
        type: "string"
      email:
        type: "string"
        format: email
        description: Unique email address for the mechanic
      phone:
        type: "string"

  ## Delete Mechanic Response
  DeleteMechanicResponse:
    type: "object"
    properties:
      message:
        type: "string"

  ## Most Serivce Tickets worked on by Mechanic List Response
...
  MechanicNameSearchResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      address:
        type: string
      email:
        type: string
      phone:
        type: string

  ####**********************************Definitions for Service_ticket Routes******************************####

  ## Create Service Ticket with Mechanic Request and Responses
...